requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
pika==1.3.2
Pillow==10.2.0
python-dotenv==1.0.1
tensorflow
ultralytics==8.1.20
